fix updater.sh
Added a condition to check if the Trash folder exists or not

+++ updater.sh
@@ -96,15 +96,17 @@ else
     printf "${RED}System not supported${NORMAL}"
 fi
 
-printf "${RED}\nShowing files in .local/share/Trash/${NORMAL}\n"
-ls -hl ~/.local/share/Trash/files
-printf "Should I clean Trash? "
-read -p "[y/n]: " ANSW 
-if [[ "$ANSW" == "y" ]]; then
-    sudo rm -rf ~/.local/share/Trash/*
-    printProgress Trash: "cleaned"
-else
-    printProgress Trash: "not cleaned"
+if [[ -d "~/.local/share/Trash/files" ]]; then
+	printf "${RED}\nShowing files in .local/share/Trash/${NORMAL}\n"
+	ls -hl ~/.local/share/Trash/files
+	printf "Should I clean Trash? "
+	read -p "[y/n]: " ANSW 
+	if [[ "$ANSW" == "y" ]]; then
+	    sudo rm -rf ~/.local/share/Trash/*
+	    printProgress Trash: "cleaned"
+	else
+	    printProgress Trash: "not cleaned"
+	fi
 fi
 
 printf "\n"
